# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Use the Pixiu APIs to your cloud",
        "title": "Pixiu API Documentation",
        "termsOfService": "https://github.com/caoyingjunz/gopixiu",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/caoyingjunz/gopixiu",
            "email": "support@gopixiu.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "paths": {
...
        "/menus": {
            "get": {
                "description": "List menus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "List menus",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "menu_type 1: 菜单,2： 按钮, 3：API,可填写多个； 默认为： 1,2,3",
                        "name": "menu_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httputils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.PageMenu"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpError"
                        }
                    }
                }
            }
        },
        "/menus/{id}": {
            "get": {
                "description": "Get menu by menu id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "Get menu by menu id",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httputils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/model.Menu"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a menu by menu id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "Update a menu by menu id",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "menu info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateMenusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpOK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete menu by menu id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "Delete menu by menu id",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpOK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpError"
                        }
                    }
                }
            }
        },
        "/menus/{id}/status/{status}": {
            "put": {
                "description": "Update a menu status by menu id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "Update a menu status by menu id",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "status ",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpOK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httputils.HttpError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "httputils.HttpOK": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "result": {
                    "type": "string",
                    "example": "any result"
                }
            }
        },
        "httputils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回的状态码",
                    "type": "integer"
                },
                "message": {
                    "description": "异常返回时的错误信息",
                    "type": "string"
                },
                "result": {
                    "description": "正常返回时的数据，可以为任意数据结构"
                }
            }
        },
        "model.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "code": {
                    "description": "前端鉴权code 例： user:role:add, user:role:delete",
                    "type": "string"
                },
                "gmt_create": {
                    "type": "string"
                },
                "gmt_modified": {
                    "type": "string"
                },
                "icon": {
                    "description": "icon",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "menu_type": {
                    "description": "菜单类型 1 左侧菜单,2 按钮, 3 非展示权限",
                    "type": "integer"
                },
                "method": {
                    "description": "操作类型 none/GET/POST/PUT/DELETE",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "integer"
                },
                "resource_version": {
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:不启用)",
                    "type": "integer"
                },
                "url": {
                    "description": "菜单URL",
                    "type": "string"
                }
            }
        },
        "model.PageMenu": {
            "type": "object",
            "properties": {
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.UpdateMenusReq": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "前端鉴权code 例： user:role:add, user:role:delete",
                    "type": "string"
                },
                "icon": {
                    "description": "icon",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "menu_type": {
                    "description": "菜单类型 1 左侧菜单,2 按钮, 3 非展示权限",
                    "type": "integer"
                },
                "method": {
                    "description": "操作类型 none/GET/POST/PUT/DELETE",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "integer"
                },
                "resource_version": {
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:不启用)",
                    "type": "integer"
                },
                "url": {
                    "description": "菜单URL",
                    "type": "string"
                }
            }
        },
...
    }
}